// src/yeying/api/user/user.ts
export enum UserStatusEnum {
  /** USER_STATUS_UNKNOWN - 未知编码，是一种占位符 */
  USER_STATUS_UNKNOWN = 0,
  /** USER_STATUS_ACTIVE - 活跃：表示用户目前正在使用软件并与其进行交互 */
  USER_STATUS_ACTIVE = 1,
  /** USER_STATUS_OFFLINE - 离线：表示用户当前不处于活跃状态，但可能会随时返回 */
  USER_STATUS_OFFLINE = 2,
  /** USER_STATUS_DISABLE - 禁用：表示用户的访问权限被暂时或永久性地限制 */
  USER_STATUS_DISABLE = 3,
  /** USER_STATUS_LOCK - 锁定：表示用户的账户被暂时性地冻结，通常是由于安全原因 */
  USER_STATUS_LOCK = 4,
  /** USER_STATUS_UNVERIFIED - 未验证：表示用户账户在创建后尚未通过验证 */
  USER_STATUS_UNVERIFIED = 5,
  /** USER_STATUS_DELETED - 删除：表示用户账户已经被永久删除，无法恢复 */
  USER_STATUS_DELETED = 6,
  /** USER_STATUS_DORMANT - 休眠：表示用户账户由于长时间未活跃而被标记为休眠状态 */
  USER_STATUS_DORMANT = 7,
  /** USER_STATUS_FREEZE - 冻结：表示用户账户在一段时间内暂时不可用 */
  USER_STATUS_FREEZE = 8,
  /** USER_STATUS_AUDIT - 待审核：表示新用户注册后正在等待管理员或审核人员审批 */
  USER_STATUS_AUDIT = 9,
  /** USER_STATUS_REFUSED - 审核未通过：表示管理员或审核人员拒绝了用户的注册请求 */
  USER_STATUS_REFUSED = 10,
  UNRECOGNIZED = -1,
}
// ...
export enum UserRoleEnum {
  /** USER_ROLE_UNKNOWN - 未知编码，是一种占位符 */
  USER_ROLE_UNKNOWN = 0,
  /** USER_ROLE_OWNER - 所有者 */
  USER_ROLE_OWNER = 1,
  /** USER_ROLE_NORMAL - 普通用户 */
  USER_ROLE_NORMAL = 2,
  UNRECOGNIZED = -1,
}
// ...
export interface UserMetadata {
  did: string;
  name: string;
  avatar: string;
  createdAt: string;
  updatedAt: string;
  signature: string;
  telephone: string;
  email: string;
}

export interface UserState {
  owner: string;
  did: string;
  role: UserRoleEnum;
  status: UserStatusEnum;
  createdAt: string;
  updatedAt: string;
  signature: string;
}

export interface AddUserRequest {
  header: MessageHeader | undefined;
  body: AddUserRequestBody | undefined;
}

export interface AddUserRequestBody {
  user: UserMetadata | undefined;
}

export interface AddUserResponse {
  header: MessageHeader | undefined;
  body: AddUserResponseBody | undefined;
}

export interface AddUserResponseBody {
  status: ResponseStatus | undefined;
  user: UserMetadata | undefined;
}

export interface UpdateUserRequest {
  header: MessageHeader | undefined;
  body: UpdateUserRequestBody | undefined;
}

export interface UpdateUserRequestBody {
  user: UserMetadata | undefined;
}

export interface UpdateUserResponse {
  header: MessageHeader | undefined;
  body: UpdateUserResponseBody | undefined;
}

export interface UpdateUserResponseBody {
  status: ResponseStatus | undefined;
  user: UserMetadata | undefined;
}

export interface DeleteUserRequest {
  header: MessageHeader | undefined;
}

export interface DeleteUserResponse {
  header: MessageHeader | undefined;
  body: DeleteUserResponseBody | undefined;
}

export interface DeleteUserResponseBody {
  status: ResponseStatus | undefined;
}

export interface UserDetailRequest {
  header: MessageHeader | undefined;
}

export interface UserDetailResponse {
  header: MessageHeader | undefined;
  body: UserDetailResponseBody | undefined;
}

export interface UserDetailResponseBody {
  status: ResponseStatus | undefined;
  detail: UserDetail | undefined;
}

export interface UserDetail {
  user: UserMetadata | undefined;
  state: UserState | undefined;
}
// ...
export const UserMetadata: MessageFns<UserMetadata> = {
  encode(message: UserMetadata, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.did !== "") {
      writer.uint32(10).string(message.did);
    }
    if (message.name !== "") {
      writer.uint32(18).string(message.name);
    }
    if (message.avatar !== "") {
      writer.uint32(26).string(message.avatar);
    }
    if (message.createdAt !== "") {
      writer.uint32(34).string(message.createdAt);
    }
    if (message.updatedAt !== "") {
      writer.uint32(42).string(message.updatedAt);
    }
    if (message.signature !== "") {
      writer.uint32(50).string(message.signature);
    }
    if (message.telephone !== "") {
      writer.uint32(58).string(message.telephone);
    }
    if (message.email !== "") {
      writer.uint32(66).string(message.email);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): UserMetadata {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseUserMetadata();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.did = reader.string();
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.name = reader.string();
          continue;
        }
        case 3: {
          if (tag !== 26) {
            break;
          }

          message.avatar = reader.string();
          continue;
        }
        case 4: {
          if (tag !== 34) {
            break;
          }

          message.createdAt = reader.string();
          continue;
        }
        case 5: {
          if (tag !== 42) {
            break;
          }

          message.updatedAt = reader.string();
          continue;
        }
        case 6: {
          if (tag !== 50) {
            break;
          }

          message.signature = reader.string();
          continue;
        }
        case 7: {
          if (tag !== 58) {
            break;
          }

          message.telephone = reader.string();
          continue;
        }
        case 8: {
          if (tag !== 66) {
            break;
          }

          message.email = reader.string();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): UserMetadata {
    return {
      did: isSet(object.did) ? globalThis.String(object.did) : "",
      name: isSet(object.name) ? globalThis.String(object.name) : "",
      avatar: isSet(object.avatar) ? globalThis.String(object.avatar) : "",
      createdAt: isSet(object.createdAt) ? globalThis.String(object.createdAt) : "",
      updatedAt: isSet(object.updatedAt) ? globalThis.String(object.updatedAt) : "",
      signature: isSet(object.signature) ? globalThis.String(object.signature) : "",
      telephone: isSet(object.telephone) ? globalThis.String(object.telephone) : "",
      email: isSet(object.email) ? globalThis.String(object.email) : "",
    };
  },

  toJSON(message: UserMetadata): unknown {
    const obj: any = {};
    if (message.did !== "") {
      obj.did = message.did;
    }
    if (message.name !== "") {
      obj.name = message.name;
    }
    if (message.avatar !== "") {
      obj.avatar = message.avatar;
    }
    if (message.createdAt !== "") {
      obj.createdAt = message.createdAt;
    }
    if (message.updatedAt !== "") {
      obj.updatedAt = message.updatedAt;
    }
    if (message.signature !== "") {
      obj.signature = message.signature;
    }
    if (message.telephone !== "") {
      obj.telephone = message.telephone;
    }
    if (message.email !== "") {
      obj.email = message.email;
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<UserMetadata>, I>>(base?: I): UserMetadata {
    return UserMetadata.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<UserMetadata>, I>>(object: I): UserMetadata {
    const message = createBaseUserMetadata();
    message.did = object.did ?? "";
    message.name = object.name ?? "";
    message.avatar = object.avatar ?? "";
    message.createdAt = object.createdAt ?? "";
    message.updatedAt = object.updatedAt ?? "";
    message.signature = object.signature ?? "";
    message.telephone = object.telephone ?? "";
    message.email = object.email ?? "";
    return message;
  },
};

function createBaseUserState(): UserState {
  return { owner: "", did: "", role: 0, status: 0, createdAt: "", updatedAt: "", signature: "" };
}

export const UserState: MessageFns<UserState> = {
  encode(message: UserState, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.owner !== "") {
      writer.uint32(10).string(message.owner);
    }
    if (message.did !== "") {
      writer.uint32(18).string(message.did);
    }
    if (message.role !== 0) {
      writer.uint32(24).int32(message.role);
    }
    if (message.status !== 0) {
      writer.uint32(32).int32(message.status);
    }
    if (message.createdAt !== "") {
      writer.uint32(42).string(message.createdAt);
    }
    if (message.updatedAt !== "") {
      writer.uint32(50).string(message.updatedAt);
    }
    if (message.signature !== "") {
      writer.uint32(58).string(message.signature);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): UserState {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseUserState();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.owner = reader.string();
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.did = reader.string();
          continue;
        }
        case 3: {
          if (tag !== 24) {
            break;
          }

          message.role = reader.int32() as any;
          continue;
        }
        case 4: {
          if (tag !== 32) {
            break;
          }

          message.status = reader.int32() as any;
          continue;
        }
        case 5: {
          if (tag !== 42) {
            break;
          }

          message.createdAt = reader.string();
          continue;
        }
        case 6: {
          if (tag !== 50) {
            break;
          }

          message.updatedAt = reader.string();
          continue;
        }
        case 7: {
          if (tag !== 58) {
            break;
          }

          message.signature = reader.string();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): UserState {
    return {
      owner: isSet(object.owner) ? globalThis.String(object.owner) : "",
      did: isSet(object.did) ? globalThis.String(object.did) : "",
      role: isSet(object.role) ? userRoleEnumFromJSON(object.role) : 0,
      status: isSet(object.status) ? userStatusEnumFromJSON(object.status) : 0,
      createdAt: isSet(object.createdAt) ? globalThis.String(object.createdAt) : "",
      updatedAt: isSet(object.updatedAt) ? globalThis.String(object.updatedAt) : "",
      signature: isSet(object.signature) ? globalThis.String(object.signature) : "",
    };
  },

  toJSON(message: UserState): unknown {
    const obj: any = {};
    if (message.owner !== "") {
      obj.owner = message.owner;
    }
    if (message.did !== "") {
      obj.did = message.did;
    }
    if (message.role !== 0) {
      obj.role = userRoleEnumToJSON(message.role);
    }
    if (message.status !== 0) {
      obj.status = userStatusEnumToJSON(message.status);
    }
    if (message.createdAt !== "") {
      obj.createdAt = message.createdAt;
    }
    if (message.updatedAt !== "") {
      obj.updatedAt = message.updatedAt;
    }
    if (message.signature !== "") {
      obj.signature = message.signature;
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<UserState>, I>>(base?: I): UserState {
    return UserState.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<UserState>, I>>(object: I): UserState {
    const message = createBaseUserState();
    message.owner = object.owner ?? "";
    message.did = object.did ?? "";
    message.role = object.role ?? 0;
    message.status = object.status ?? 0;
    message.createdAt = object.createdAt ?? "";
    message.updatedAt = object.updatedAt ?? "";
    message.signature = object.signature ?? "";
    return message;
  },
};
// ...
type Builtin = Date | Function | Uint8Array | string | number | boolean | undefined;

export type DeepPartial<T> = T extends Builtin ? T
  : T extends globalThis.Array<infer U> ? globalThis.Array<DeepPartial<U>>
  : T extends ReadonlyArray<infer U> ? ReadonlyArray<DeepPartial<U>>
  : T extends { $case: string } ? { [K in keyof Omit<T, "$case">]?: DeepPartial<T[K]> } & { $case: T["$case"] }
  : T extends {} ? { [K in keyof T]?: DeepPartial<T[K]> }
  : Partial<T>;

type KeysOfUnion<T> = T extends T ? keyof T : never;
export type Exact<P, I extends P> = P extends Builtin ? P
  : P & { [K in keyof P]: Exact<P[K], I[K]> } & { [K in Exclude<keyof I, KeysOfUnion<P>>]: never };
// ...
export interface MessageFns<T> {
  encode(message: T, writer?: BinaryWriter): BinaryWriter;
  decode(input: BinaryReader | Uint8Array, length?: number): T;
  fromJSON(object: any): T;
  toJSON(message: T): unknown;
  create<I extends Exact<DeepPartial<T>, I>>(base?: I): T;
  fromPartial<I extends Exact<DeepPartial<T>, I>>(object: I): T;
}